requirements.txt
feedparser>=6.0.10
Mastodon.py>=1.5.1
requests>=2.31.0
